Implement short and long form download table
Strategy to avoid code duplication was to use a partial that builds
both tables, but toggle set to show only one of them.

JSON file with data for the tables can be downloaded. Need 2 copies
of the JSON file because the can't use static files in partials, needs to be in Site.Data to be accessible.
CSV file to be added later.